db 변경

+++ src/models/Menu.ts
@@ -35,13 +35,13 @@ export class Menu extends Model<MenuAttributes, MenuCreationAttributes> implemen
             {
                 seq: {
                     autoIncrement: true,
-                    type: DataTypes.BIGINT.UNSIGNED,
+                    type: DataTypes.SMALLINT.UNSIGNED,
                     allowNull: false,
                     primaryKey: true,
                     comment: '메뉴 Seq',
                 },
                 ctgSeq: {
-                    type: DataTypes.BIGINT.UNSIGNED,
+                    type: DataTypes.SMALLINT.UNSIGNED,
                     allowNull: false,
                     comment: '메뉴 카테고리 Seq',
                 },
